requirements.txt
# =============================================================================
# The Architect - Requirements
# Python 3.10+ | CUDA 11.8+ recommended
# =============================================================================

# -----------------------------------------------------------------------------
# Core Deep Learning
# -----------------------------------------------------------------------------
torch>=2.0.0,<2.2.0
torchvision>=0.15.0,<0.17.0
torchaudio>=2.0.0

# -----------------------------------------------------------------------------
# Vision-Language Models
# -----------------------------------------------------------------------------
open-clip-torch>=2.20.0,<2.25.0
transformers>=4.30.0,<4.40.0
timm>=0.9.0

# -----------------------------------------------------------------------------
# Object Detection
# -----------------------------------------------------------------------------
# YOLOv8 - 主要检测器，轻量且易安装
ultralytics>=8.0.180,<8.2.0

# [可选] Grounding DINO - 更强的零样本检测能力
# 注意：安装较复杂，Windows 需特殊处理
# pip install groundingdino-py@git+https://github.com/IDEA-Research/GroundingDINO.git

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
numpy>=1.24.0,<2.0.0
pillow>=10.0.0
opencv-python>=4.8.0,<4.9.0
pycocotools>=2.0.6
h5py>=3.9.0

# COCO API 依赖
matplotlib>=3.7.0
scikit-image>=0.21.0

# -----------------------------------------------------------------------------
# Visualization
# -----------------------------------------------------------------------------
matplotlib>=3.7.0,<3.9.0
seaborn>=0.12.0,<0.14.0
plotly>=5.15.0

# TensorBoard 日志
tensorboard>=2.13.0,<2.16.0
tensorboardX>=2.6.0

# [可选] Weights & Biases 实验跟踪
wandb>=0.15.0,<0.17.0

# -----------------------------------------------------------------------------
# Configuration & Utilities
# -----------------------------------------------------------------------------
pyyaml>=6.0
omegaconf>=2.3.0
tqdm>=4.65.0
einops>=0.7.0
safetensors>=0.3.0

# -----------------------------------------------------------------------------
# Training Optimizations
# -----------------------------------------------------------------------------
accelerate>=0.21.0,<0.26.0

# [可选] DeepSpeed 大规模训练
# deepspeed>=0.10.0

# -----------------------------------------------------------------------------
# Evaluation & Analysis
# -----------------------------------------------------------------------------
scikit-learn>=1.3.0,<1.5.0
scipy>=1.11.0,<1.13.0
pandas>=2.0.0

# -----------------------------------------------------------------------------
# Model Hub (China Mirror Support)
# -----------------------------------------------------------------------------
modelscope>=1.10.0

# Hugging Face 生态
datasets>=2.14.0
tokenizers>=0.13.0

# -----------------------------------------------------------------------------
# Jupyter & Development
# -----------------------------------------------------------------------------
jupyter>=1.0.0
ipywidgets>=8.0.0
ipython>=8.14.0

# 代码格式化
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# -----------------------------------------------------------------------------
# System Utilities
# -----------------------------------------------------------------------------
psutil>=5.9.0
packaging>=23.0

# =============================================================================
# Installation Notes
# =============================================================================
# 
# 1. 推荐安装命令:
#    pip install -r requirements.txt
#
# 2. 如果使用 Conda:
#    conda install pytorch torchvision torchaudio pytorch-cuda=11.8 -c pytorch -c nvidia
#    pip install -r requirements.txt
#
# 3. 国内用户加速 (使用清华镜像):
#    pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
#
# 4. 仅安装必要依赖 (最小化安装):
#    pip install torch torchvision open-clip-torch ultralytics pyyaml matplotlib
#
# 5. 验证安装:
#    python scripts/test_setup.py
# =============================================================================
